[REFACTOR] (1) 패키지 수정 (2) ResponseDto에 @JsonIgnoreProperties 추가
