--- src/main/java/com/spring/tming/domain/applicant/ApplicantSaveRes.java
@@ -1,3 +1,0 @@
-package com.spring.tming.domain.applicant;
-
-public class ApplicantSaveRes {}
