+++ src/main/java/com/spring/tming/domain/applicant/dto/request/ApplicantSaveReq.java
@@ -1,6 +1,6 @@
-package com.spring.tming.domain.applicant;
+package com.spring.tming.domain.applicant.dto.request;
 
-import java.io.Serializable;
+import com.spring.tming.global.meta.Job;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -8,14 +8,15 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ApplicantId implements Serializable {
-
+public class ApplicantSaveReq {
     private Long postId;
+    private Job job;
     private Long userId;
 
     @Builder
-    private ApplicantId(Long postId, Long userId) {
+    private ApplicantSaveReq(Long postId, Job job, Long userId) {
         this.postId = postId;
+        this.job = job;
         this.userId = userId;
     }
 }
