+++ src/main/java/com/spring/tming/domain/applicant/dto/response/ApplicantSaveRes.java
@@ -1,0 +1,6 @@
+package com.spring.tming.domain.applicant.dto.response;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+@JsonIgnoreProperties
+public class ApplicantSaveRes {}
